drm-shim/v3d: mem_size is now an uint32_t
Equivalent to:
commit 646a8338b19182a73c0fb31a5e4bb845bf611f2e
Author: Alejandro Piñeiro <apinheiro@igalia.com>
Date:   Fri Apr 23 12:12:49 2021 +0200

    v3d/simulator: hw mem is now an v3d_size_t, typedef to uint32_t

+++ src/broadcom/drm-shim/v3d.h
@@ -38,7 +38,7 @@ struct v3d_shim_device {
         /* Base hardware address of the heap. */
         uint32_t mem_base;
         /* Size of the heap. */
-        size_t mem_size;
+        uint32_t mem_size;
 
         /* Allocator for the GPU virtual addresses. */
         struct util_vma_heap heap;
